+++ packages/viewer/web/src/List.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { getBlockIndex } from './utils/draftUtils';
-import { hasText, safariOrFirefox } from './utils/textUtils';
+import { hasText } from './utils/textUtils';
 import { isPaywallSeo, getPaywallSeoClass } from './utils/paywallSeo';
 import { getDirectionFromAlignmentAndTextDirection } from 'wix-rich-content-common';
 import { getInteractionWrapper, DefaultInteractionWrapper } from './utils/getInteractionWrapper';
@@ -52,11 +52,10 @@ const List = ({
         let paragraphGroup = [];
         const result = [];
         const textClassName = getBlockStyleClasses(dataEntry, mergedStyles, textDirection);
-        const safariOrFirefoxJustify =
-          dataEntry?.textAlignment === 'justify' && safariOrFirefox() && hasText(children);
+        const hasJustifyText = dataEntry?.textAlignment === 'justify' && hasText(children);
         const elementProps = key => ({
           className: classNames(mergedStyles.elementSpacing, textClassName, {
-            [styles.hasTextAndSafariOrFirefox]: safariOrFirefoxJustify,
+            [styles.hasJustifyText]: hasJustifyText,
           }),
           key,
         });